BE/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
algokit-utils==4.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bases==0.3.0
brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
copier==9.7.1
cryptography==44.0.3
cssselect2==0.8.0
deprecation==2.1.0
distro==1.9.0
dunamai==1.24.1
elevenlabs==2.3.0
faker==37.4.2
fastapi==0.115.12
fonttools==4.58.5
frozenlist==1.7.0
funcy==2.0
google-auth==2.40.3
google-genai==1.23.0
gotrue==2.12.2
grpclib==0.4.7
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.2.1
jinja2==3.1.6
jinja2-ansible-filters==1.3.2
jiter==0.10.0
jsondiff==2.2.1
keyring==25.6.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
markupsafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
modal==1.0.4
more-itertools==10.7.0
msgpack==1.1.1
mslex==1.3.0
multidict==6.4.4
multiformats==0.3.1
multiformats-config==0.3.1
numpy==2.3.1
openai==1.86.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
pip==25.1.1
pipdeptree==2.27.0
platformdirs==4.3.8
pluggy==1.6.0
plumbum==1.9.0
postgrest==1.1.1
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
py-algorand-sdk==2.9.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyclip==0.7.0
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pydyf==0.11.0
pygments==2.19.1
pyjwt==2.10.1
pynacl==1.5.0
pyphen==0.17.2
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
questionary==1.10.0
realtime==2.5.3
requests==2.32.4
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
storage3==0.12.0
strenum==0.4.15
supabase==2.16.0
supafunc==0.10.1
synchronicity==0.9.15
tenacity==8.5.0
textual==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tqdm==4.67.1
twilio==9.6.5
typer==0.16.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-extensions==4.14.0
typing-inspection==0.4.1
typing-validation==1.2.12
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.4.0
uuid==1.30
uvicorn==0.34.3
watchfiles==1.0.5
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
yarl==1.20.1
zopfli==0.2.3.post1
